xen/x86: report domain id on cpuid
Use the ECX register of the hypervisor leaf 5. The EAX register on
this leaf is a flags field that can be used to notice the presence of
the domain id in ECX. Note that this is only available to HVM guests.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
master commit: 021e1b7c7a70c71f09c944d3e539dbf476fe48d9
master date: 2018-01-16 18:34:04 +0000